+++ src/main/java/myorg/util/ScoreStruct.java
@@ -2,8 +2,8 @@ package myorg.util;
 
 public class ScoreStruct implements Comparable<ScoreStruct> {
     public float predict = 0.0f;
-    public int positive = 0;
-    public int negative = 0;
+    public float positive = 0;
+    public float negative = 0;
 
     public int compareTo(ScoreStruct o) {
         float diff = this.predict - o.predict;
